Crate ruma::api::client

(De)serializable types for the Matrix Client-Server API. These types can be shared by client and server code.

Modules

account

Endpoints for account registration and management.

alias

Endpoints for room aliases.

appservice

Endpoints part of the application service extension of the client-server API

backup

Endpoints for server-side key backups.

config

Endpoints for client configuration.

context

Endpoints for event context.

device

Endpoints for managing devices.

directory

Endpoints for the public room directory.

discovery

Server discovery endpoints.

error

Errors that can be sent from the homeserver.

filter

Endpoints for event filters.

keys

Endpoints for key management

knock

Endpoints to knock on a room.

media

Endpoints for the media repository.

membership

Endpoints for room membership.

message

Endpoints for sending and receiving messages

presence

Endpoints for user presence.

profile

Endpoints for user profiles.

push

Endpoints for push notifications.

read_marker

Endpoints for “fully read markers”

receipt

Endpoints for event receipts.

redact

Endpoints for event redaction.

room

Endpoints for room management.

search

Endpoints for event searches.

server

Endpoints for server administration.

session

Endpoints for user session management.

space

Endpoints for spaces.

state

Endpoints for managing room state

sync

Endpoints for getting and synchronizing events.

tag

Endpoints for tagging rooms.

thirdparty

Endpoints for third party lookups

to_device

Endpoints for client devices to exchange information not persisted in room DAG.

typing

Endpoints for typing notifications.

uiaa

Module for User-Interactive Authentication API types.

user_directory

Endpoints for the user directory.

voip

Endpoints for Voice over IP.

Structs

Error

A Matrix Error