Struct ruma::api::client::r0::keys::upload_keys::Request

#[non_exhaustive]pub struct Request {
    pub device_keys: Option<DeviceKeys>,
    pub one_time_keys: Option<BTreeMap<DeviceKeyId, OneTimeKey>>,
}

This is supported on crate feature api and (crate features client-api or client-api-c or client-api-s) only.

Data for a request to the upload_keys API endpoint.

Publishes end-to-end encryption keys for the device.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
device_keys: Option<DeviceKeys>

Identity keys for the device. May be absent if no new identity keys are required.

one_time_keys: Option<BTreeMap<DeviceKeyId, OneTimeKey>>

One-time public keys for “pre-key” messages.

Implementations

impl Request

pub fn new() -> Request

Creates an empty Request.