Module ruma::api::client::r0::search::search_events

This is supported on crate features client-api or client-api-c or client-api-s only.

POST /_matrix/client/r0/search

Structs

Categories

Categories of events that can be searched for.

Criteria

Criteria for searching a category of events.

EventContext

Configures whether any context for the events returned are included in the response.

EventContextResult

Context for search results, if requested.

Grouping

A grouping for partitioning the result set.

Groupings

Requests that the server partitions the result set based on the provided list of keys.

IncomingCategories

‘Incoming’ variant of Categories.

IncomingCriteria

‘Incoming’ variant of Criteria.

IncomingGroupings

‘Incoming’ variant of Groupings.

IncomingRequest

‘Incoming’ variant of Request.

Request

Data for a request to the search API endpoint.

Response

Data in the response from the search API endpoint.

ResultCategories

Categories of events that can be searched for.

ResultGroup

A grouping of results, if requested.

ResultRoomEvents

Categories of events that can be searched for.

SearchResult

A search result.

UserProfile

A user profile.

Enums

GroupingKey

The key within events to use for this grouping.

OrderBy

The order in which to search for results.

RoomIdOrUserId

Represents either a room or user ID for returning grouped search results.

SearchKeys

The keys to search for.

Constants

METADATA

Metadata for the search API endpoint.