Enum ruma::api::client::room::Visibility

#[non_exhaustive]
pub enum Visibility {
    Public,
    Private,
}

Whether or not a newly created room will be listed in the room directory.

This type can hold an arbitrary string. To build this with a custom value, convert it from a string with ::from() / .into(). To check for values that are not available as a documented variant here, use its string representation, obtained through .as_str().

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Public

Indicates that the room will be shown in the published room list.

Private

Indicates that the room will not be shown in the published room list.

Implementations

impl Visibility

pub fn as_str(&self) -> &str

Creates a string slice from this Visibility.