Struct ruma::api::client::to_device::send_event_to_device::v3::Request

#[non_exhaustive]
pub struct Request<'a> {
    pub event_type: &'a str,
    pub txn_id: &'a TransactionId,
    pub messages: BTreeMap<OwnedUserId, BTreeMap<DeviceIdOrAllDevices, Raw<AnyToDeviceEventContent>, Global>, Global>,
}

Data for a request to the send_event_to_device API endpoint.

Send an event to a device or devices.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

event_type: &'a str

Type of event being sent to each device.

txn_id: &'a TransactionId

A request identifier unique to the access token used to send the request.

messages: BTreeMap<OwnedUserId, BTreeMap<DeviceIdOrAllDevices, Raw<AnyToDeviceEventContent>, Global>, Global>

Messages to send.

Different message events can be sent to different devices in the same request, but all events within one request must be of the same type.

Implementations

impl<'a> Request<'a>

pub fn new_raw(
event_type: &'a str,
txn_id: &'a TransactionId,
messages: BTreeMap<OwnedUserId, BTreeMap<DeviceIdOrAllDevices, Raw<AnyToDeviceEventContent>, Global>, Global>
) -> Request<'a>

Creates a new Request with the given event type, transaction ID and raw messages.