Module ruma::api::error

This module contains types for all kinds of errors that can occur when converting between http requests / responses and ruma’s representation of matrix API requests / responses.

Structs

MatrixError

A general-purpose Matrix error type consisting of an HTTP status code and a JSON body.

UnknownVersionError

An error that happens when Ruma cannot understand a Matrix version.

Enums

DeserializationError

An error when converting a http request / response to one of ruma’s endpoint-specific request / response types.

FromHttpRequestError

An error when converting a http request to one of ruma’s endpoint-specific request types.

FromHttpResponseError

An error when converting a http response to one of Ruma’s endpoint-specific response types.

HeaderDeserializationError

An error with the http headers.

IntoHttpError

An error when converting one of ruma’s endpoint-specific request or response types to the corresponding http type.

ServerError

An error was reported by the server (HTTP status code 4xx or 5xx)