Module ruma::api::federation::transactions::send_transaction_message

This is supported on crate feature api and (crate features federation-api or federation-api-c or federation-api-s) only.

Endpoint to send live activity messages to another server.

Modules

v1

PUT /_matrix/federation/v1/send/{txnId}