Crate ruma::api::identity_service

This is supported on crate features identity-service-api or identity-service-api-c or identity-service-api-s only.

(De)serializable types for the Matrix Identity Service API. These types can be shared by client and identity service code.

Modules

association

Endpoints to create associations with a Matrix ID on the identity server.

authentication

Endpoints to authenticate with an identity server.

invitation

Endpoints to store and sign invitations.

keys

Endpoints to retrieve, update, and validate keys with an identity server.

lookup

Endpoints to look up Matrix IDs bound to 3PIDs.

status

Endpoint to check the status of an identity server.

tos

Endpoints to retrieve and accept terms of service of an identity server.