Module ruma::common::events::audio

This is supported on crate features events and unstable-msc3246 only.

Types for extensible audio message events (MSC3246).

Structs

Amplitude

The amplitude of a waveform sample.

AudioContent

Audio content.

AudioEventContent

The payload for an extensible audio message.

RedactedAudioEventContent

Redacted form of AudioEventContent

Waveform

The waveform representation of audio content.

Enums

WaveformError

An error encountered when trying to convert to a Waveform.

Type Definitions

AudioEvent

An m.audio event.

RedactedAudioEvent

An m.audio event that has been redacted.

RedactedSyncAudioEvent

An m.audio event from a sync_events response that has been redacted.

SyncAudioEvent

An m.audio event from a sync_events response.