Enum ruma::events::AnySyncMessageLikeEvent

#[non_exhaustive]
pub enum AnySyncMessageLikeEvent {

    CallAnswer(SyncMessageLikeEvent<CallAnswerEventContent>),
    CallInvite(SyncMessageLikeEvent<CallInviteEventContent>),
    CallHangup(SyncMessageLikeEvent<CallHangupEventContent>),
    CallCandidates(SyncMessageLikeEvent<CallCandidatesEventContent>),
    KeyVerificationReady(SyncMessageLikeEvent<KeyVerificationReadyEventContent>),
    KeyVerificationStart(SyncMessageLikeEvent<KeyVerificationStartEventContent>),
    KeyVerificationCancel(SyncMessageLikeEvent<KeyVerificationCancelEventContent>),
    KeyVerificationAccept(SyncMessageLikeEvent<KeyVerificationAcceptEventContent>),
    KeyVerificationKey(SyncMessageLikeEvent<KeyVerificationKeyEventContent>),
    KeyVerificationMac(SyncMessageLikeEvent<KeyVerificationMacEventContent>),
    KeyVerificationDone(SyncMessageLikeEvent<KeyVerificationDoneEventContent>),
    Reaction(SyncMessageLikeEvent<ReactionEventContent>),
    RoomEncrypted(SyncMessageLikeEvent<RoomEncryptedEventContent>),
    RoomMessage(SyncMessageLikeEvent<RoomMessageEventContent>),
    RoomMessageFeedback(SyncMessageLikeEvent<RoomMessageFeedbackEventContent>),
    RoomRedaction(SyncRoomRedactionEvent),
    Sticker(SyncMessageLikeEvent<StickerEventContent>),
    // some variants omitted

}

Any message-like event.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CallAnswer(SyncMessageLikeEvent<CallAnswerEventContent>)

m.call.answer

CallInvite(SyncMessageLikeEvent<CallInviteEventContent>)

m.call.invite

CallHangup(SyncMessageLikeEvent<CallHangupEventContent>)

m.call.hangup

CallCandidates(SyncMessageLikeEvent<CallCandidatesEventContent>)

m.call.candidates

KeyVerificationReady(SyncMessageLikeEvent<KeyVerificationReadyEventContent>)

m.key.verification.ready

KeyVerificationStart(SyncMessageLikeEvent<KeyVerificationStartEventContent>)

m.key.verification.start

KeyVerificationCancel(SyncMessageLikeEvent<KeyVerificationCancelEventContent>)

m.key.verification.cancel

KeyVerificationAccept(SyncMessageLikeEvent<KeyVerificationAcceptEventContent>)

m.key.verification.accept

KeyVerificationKey(SyncMessageLikeEvent<KeyVerificationKeyEventContent>)

m.key.verification.key

KeyVerificationMac(SyncMessageLikeEvent<KeyVerificationMacEventContent>)

m.key.verification.mac

KeyVerificationDone(SyncMessageLikeEvent<KeyVerificationDoneEventContent>)

m.key.verification.done

Reaction(SyncMessageLikeEvent<ReactionEventContent>)

m.reaction

RoomEncrypted(SyncMessageLikeEvent<RoomEncryptedEventContent>)

m.room.encrypted

RoomMessage(SyncMessageLikeEvent<RoomMessageEventContent>)

m.room.message

RoomMessageFeedback(SyncMessageLikeEvent<RoomMessageFeedbackEventContent>)

m.room.message.feedback

RoomRedaction(SyncRoomRedactionEvent)

m.room.redaction

Sticker(SyncMessageLikeEvent<StickerEventContent>)

m.sticker

Implementations

impl AnySyncMessageLikeEvent

pub fn event_type(&self) -> MessageLikeEventType

Returns the type of this event.

pub fn original_content(&self) -> Option<AnyMessageLikeEventContent>

Returns the content for this event if it is not redacted, or None if it is.

pub fn origin_server_ts(&self) -> MilliSecondsSinceUnixEpoch

Returns this event’s origin_server_ts field.

pub fn event_id(&self) -> &EventId

Returns this event’s event_id field.

pub fn sender(&self) -> &UserId

Returns this event’s sender field.

pub fn transaction_id(&self) -> Option<&TransactionId>

Returns this event’s transaction_id from inside unsigned, if there is one.

impl AnySyncMessageLikeEvent

pub fn into_full_event(self, room_id: OwnedRoomId) -> AnyMessageLikeEvent

Convert this sync event into a full event (one with a room_id field).