Enum ruma::events::StateEventType

#[non_exhaustive]
pub enum StateEventType {

    PolicyRuleRoom,
    PolicyRuleServer,
    PolicyRuleUser,
    RoomAliases,
    RoomAvatar,
    RoomCanonicalAlias,
    RoomCreate,
    RoomEncryption,
    RoomGuestAccess,
    RoomHistoryVisibility,
    RoomJoinRules,
    RoomMember,
    RoomName,
    RoomPinnedEvents,
    RoomPowerLevels,
    RoomServerAcl,
    RoomThirdPartyInvite,
    RoomTombstone,
    RoomTopic,
    SpaceChild,
    SpaceParent,
    CallMember,

}

The type of StateEvent this is.

This type can hold an arbitrary string. To build events with a custom type, convert it from a string with ::from() / .into(). To check for events that are not available as a documented variant here, use its string representation, obtained through .to_string().

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

PolicyRuleRoom

m.policy.rule.room

PolicyRuleServer

m.policy.rule.server

PolicyRuleUser

m.policy.rule.user

RoomAliases

m.room.aliases

RoomAvatar

m.room.avatar

RoomCanonicalAlias

m.room.canonical_alias

RoomCreate

m.room.create

RoomEncryption

m.room.encryption

RoomGuestAccess

m.room.guest_access

RoomHistoryVisibility

m.room.history_visibility

RoomJoinRules

m.room.join_rules

RoomMember

m.room.member

RoomName

m.room.name

RoomPinnedEvents

m.room.pinned_events

RoomPowerLevels

m.room.power_levels

RoomServerAcl

m.room.server_acl

RoomThirdPartyInvite

m.room.third_party_invite

RoomTombstone

m.room.tombstone

RoomTopic

m.room.topic

SpaceChild

m.space.child

SpaceParent

m.space.parent

CallMember

org.matrix.msc3401.call.member