Module ruma::events::file

This is supported on crate features events and unstable-msc3551 only.

Types for extensible file message events (MSC3551).

Structs

EncryptedContent

The encryption info of a file sent to a room with end-to-end encryption enabled.

EncryptedContentInit

Initial set of fields of EncryptedContent.

FileContent

File content.

FileContentInfo

Information about a file content.

FileEventContent

The payload for an extensible file message.

RedactedFileEventContent

Redacted form of FileEventContent

Type Definitions

FileEvent

An m.file event.

OriginalFileEvent

An m.file event.

OriginalSyncFileEvent

An m.file event from a sync_events response.

RedactedFileEvent

An m.file event that has been redacted.

RedactedSyncFileEvent

An m.file event from a sync_events response that has been redacted.

SyncFileEvent

An m.file event from a sync_events response.