Module ruma::events::relation

Available on crate feature events only.

Types describing relationships between events.

Structs

AnnotationChunkunstable-msc2677

The first chunk of annotations with a token for loading more.

BundledAnnotationunstable-msc2677

Summary of all annotations to an event with the given key and type.

BundledReplacementunstable-msc2676

A bundled replacement.

BundledThreadunstable-msc3440

A bundled thread.

Relations

Bundled aggregations of related child events.

Enums

AnnotationTypeunstable-msc2677

Type of annotation.

RelationType

Relation types as defined in rel_type of an m.relates_to field.