Struct ruma::events::room::create::RoomCreateEventContent

#[non_exhaustive]
pub struct RoomCreateEventContent {
    pub creator: Option<OwnedUserId>,
    pub federate: bool,
    pub room_version: RoomVersionId,
    pub predecessor: Option<PreviousRoom>,
    pub room_type: Option<RoomType>,
}

The content of an m.room.create event.

This is the first event in a room and cannot be changed.

It acts as the root of all other events.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

creator: Option<OwnedUserId>
👎Deprecated: Since Matrix 1.8. This field was removed in Room version 11, clients should use the event’s sender instead

The user_id of the room creator.

This is set by the homeserver.

This is required in room versions 1 trough 10, but is removed starting from room version 11.

federate: bool

Whether or not this room’s data should be transferred to other homeservers.

room_version: RoomVersionId

The version of the room.

Defaults to RoomVersionId::V1.

predecessor: Option<PreviousRoom>

A reference to the room this room replaces, if the previous room was upgraded.

room_type: Option<RoomType>

The room type.

This is currently only used for spaces.

Implementations

impl RoomCreateEventContent

pub fn new_v1(creator: OwnedUserId) -> RoomCreateEventContent

Creates a new RoomCreateEventContent with the given creator, as required for room versions 1 through 10.

pub fn new_v11() -> RoomCreateEventContent

Creates a new RoomCreateEventContent with the default values and no creator, as introduced in room version 11.

The room version is set to RoomVersionId::V11.