Type Alias ruma::events::room::create::PossiblyRedactedRoomCreateEventContent

pub type PossiblyRedactedRoomCreateEventContent = RoomCreateEventContent;

The possibly redacted form of RoomCreateEventContent.

This type is used when it’s not obvious whether the content is redacted or not.

Aliased Type

struct PossiblyRedactedRoomCreateEventContent {
    pub creator: Option<OwnedUserId>,
    pub federate: bool,
    pub room_version: RoomVersionId,
    pub predecessor: Option<PreviousRoom>,
    pub room_type: Option<RoomType>,
}

Fields
creator: Option<OwnedUserId>
👎Deprecated: Since Matrix 1.8. This field was removed in Room version 11, clients should use the event’s sender instead

The user_id of the room creator.

This is set by the homeserver.

This is required in room versions 1 trough 10, but is removed starting from room version 11.

federate: bool

Whether or not this room’s data should be transferred to other homeservers.

room_version: RoomVersionId

The version of the room.

Defaults to RoomVersionId::V1.

predecessor: Option<PreviousRoom>

A reference to the room this room replaces, if the previous room was upgraded.

room_type: Option<RoomType>

The room type.

This is currently only used for spaces.