Module ruma::events::room::join_rules

Types for the m.room.join_rules event.

Structs

RedactedRoomJoinRulesEventContent

Redacted form of RoomJoinRulesEventContent

Restricted

Configuration of the Restricted join rule.

RoomJoinRulesEventContent

The content of an m.room.join_rules event.

RoomMembership

Allow rule which grants permission to join based on the membership of another room.

Enums

AllowRule

An allow rule which defines a condition that allows joining a room.

JoinRule

The rule used for users wishing to join this room.

Type Definitions

InitialRoomJoinRulesEvent

An m.room.join_rules event for creating a room.

RedactedRoomJoinRulesEvent

An m.room.join_rules event that has been redacted.

RedactedSyncRoomJoinRulesEvent

An m.room.join_rules event from a sync_events response that has been redacted.

RoomJoinRulesEvent

An m.room.join_rules event.

StrippedRoomJoinRulesEvent

An m.room.join_rules event from an invited room preview.

SyncRoomJoinRulesEvent

An m.room.join_rules event from a sync_events response.