Module ruma::events::room::message

Types for the m.room.message event.

Modules

sanitize

Convenience methods and types to sanitize text messages.

Structs

AudioInfo

Metadata about an audio clip.

AudioMessageEventContent

The payload for an audio message.

EmoteMessageEventContent

The payload for an emote message.

FileInfo

Metadata about a file.

FileMessageEventContent

The payload for a file message.

FormattedBody

Common message event content fields for message types that have separate plain-text and formatted representations.

ImageMessageEventContent

The payload for an image message.

KeyVerificationRequestEventContent

The payload for a key verification request message.

LocationInfo

Thumbnail info associated with a location.

LocationMessageEventContent

The payload for a location message.

NoticeMessageEventContent

The payload for a notice message.

RedactedRoomMessageEventContent

Redacted form of RoomMessageEventContent

ReplacementMetadata

Metadata about an event to be replaced.

RoomMessageEventContent

The content of an m.room.message event.

RoomMessageEventContentWithoutRelation

Form of RoomMessageEventContent without relation.

ServerNoticeMessageEventContent

The payload for a server notice message.

TextMessageEventContent

The payload for a text message.

UnstableAmplitude

The unstable version of the amplitude of a waveform sample.

UnstableAudioDetailsContentBlock

Extensible event fallback data for audio messages, from the first version of MSC3245.

UnstableVoiceContentBlock

Extensible event fallback data for voice messages, from the first version of MSC3245.

VideoInfo

Metadata about a video.

VideoMessageEventContent

The payload for a video message.

Enums

AddMentions

Whether or not to add intentional Mentions when sending a reply.

ForwardThread

Whether or not to forward a Relation::Thread when sending a reply.

LimitType

Types of usage limits.

MessageFormat

The format for the formatted representation of a message body.

MessageType

The content that is specific to each message type variant.

Relation

Message event relationship.

RelationWithoutReplacement

Message event relationship, except a replacement.

ReplyWithinThread

Whether or not the message is a reply inside a thread.

ServerNoticeType

Types of server notices.

Functions

deserialize_relation

Deserialize an event’s relates_to field.

Type Aliases

OriginalRoomMessageEvent

An m.room.message event.

OriginalSyncRoomMessageEvent

An m.room.message event from a sync_events response.

RedactedRoomMessageEvent

An m.room.message event that has been redacted.

RedactedSyncRoomMessageEvent

An m.room.message event from a sync_events response that has been redacted.

RoomMessageEvent

An m.room.message event.

SyncRoomMessageEvent

An m.room.message event from a sync_events response.