Enum ruma::events::room::message::MessageType

pub enum MessageType {
    Audio(AudioMessageEventContent),
    Emote(EmoteMessageEventContent),
    File(FileMessageEventContent),
    Image(ImageMessageEventContent),
    Location(LocationMessageEventContent),
    Notice(NoticeMessageEventContent),
    ServerNotice(ServerNoticeMessageEventContent),
    Text(TextMessageEventContent),
    Video(VideoMessageEventContent),
    VerificationRequest(KeyVerificationRequestEventContent),
    // some variants omitted
}

The content that is specific to each message type variant.

Variants

Audio(AudioMessageEventContent)

Tuple Fields

0: AudioMessageEventContent

An audio message.

Emote(EmoteMessageEventContent)

Tuple Fields

0: EmoteMessageEventContent

An emote message.

File(FileMessageEventContent)

Tuple Fields

0: FileMessageEventContent

A file message.

Image(ImageMessageEventContent)

Tuple Fields

0: ImageMessageEventContent

An image message.

Location(LocationMessageEventContent)

Tuple Fields

0: LocationMessageEventContent

A location message.

Notice(NoticeMessageEventContent)

Tuple Fields

0: NoticeMessageEventContent

A notice message.

ServerNotice(ServerNoticeMessageEventContent)

Tuple Fields

0: ServerNoticeMessageEventContent

A server notice message.

Text(TextMessageEventContent)

Tuple Fields

0: TextMessageEventContent

A text message.

Video(VideoMessageEventContent)

Tuple Fields

0: VideoMessageEventContent

A video message.

VerificationRequest(KeyVerificationRequestEventContent)

Tuple Fields

0: KeyVerificationRequestEventContent

A request to initiate a key verification.

Implementations

impl MessageType

pub fn new(
msgtype: &str,
body: String,
data: Map<String, Value>
) -> Result<MessageType, Error>

Creates a new MessageType with the given msgtype string and data.

Prefer to use the public variants of MessageType where possible; this constructor is meant be used for unsupported message types only and does not allow setting arbitrary data for supported ones.

Errors

Returns an error if the msgtype is known and serialization of data to the corresponding MessageType variant fails.

pub fn msgtype(&self) -> &str

Returns a reference to the msgtype string.

pub fn body(&self) -> &str

Return a reference to the message body.

pub fn data(&self) -> Cow<'_, Map<String, Value>>

Returns the associated data.

The returned JSON object won’t contain the msgtype and body fields, use .msgtype() / .body() to access those.

Prefer to use the public variants of MessageType where possible; this method is meant to be used for custom message types only.