Function ruma::events::room::message::sanitize::sanitize_html

pub fn sanitize_html(
    s: &str, 
    mode: HtmlSanitizerMode, 
    remove_reply_fallback: RemoveReplyFallback
) -> String

Available on crate features events and unstable-sanitize only.

Sanitize the given HTML string.

This removes the tags and attributes that are not listed in the Matrix specification.

It can also optionally remove the rich reply fallback.