Module ruma::events::room::power_levels

Types for the m.room.power_levels event.

Structs

RedactedRoomPowerLevelsEventContent

Redacted form of RoomPowerLevelsEventContent.

RoomPowerLevels

The effective power levels of a room.

RoomPowerLevelsEventContent

The content of an m.room.power_levels event.

Enums

NotificationPowerLevelType

The notification types that can be limited by power levels.

PowerLevelAction

The actions that can be limited by power levels.

Type Definitions

InitialRoomPowerLevelsEvent

An m.room.power_levels event for creating a room.

OriginalRoomPowerLevelsEvent

An m.room.power_levels event.

OriginalSyncRoomPowerLevelsEvent

An m.room.power_levels event from a sync_events response.

PossiblyRedactedRoomPowerLevelsEventContent

The possibly redacted form of RoomPowerLevelsEventContent.

RedactedRoomPowerLevelsEvent

An m.room.power_levels event that has been redacted.

RedactedSyncRoomPowerLevelsEvent

An m.room.power_levels event from a sync_events response that has been redacted.

RoomPowerLevelsEvent

An m.room.power_levels event.

StrippedRoomPowerLevelsEvent

An m.room.power_levels event from an invited room preview.

SyncRoomPowerLevelsEvent

An m.room.power_levels event from a sync_events response.