Trait ruma::events::Redact

pub trait Redact {
    type Redacted;
    fn redact(
        self, 
        redaction: SyncRedactionEvent, 
        version: &RoomVersionId
    ) -> Self::Redacted;
}

This is supported on crate feature events only.

Trait to define the behavior of redacting an event.

Associated Types

type Redacted

The redacted form of the event.

Required methods

fn redact(
self,
redaction: SyncRedactionEvent,
version: &RoomVersionId
) -> Self::Redacted

Transforms self into a redacted form (removing most fields) according to the spec.

A small number of events have room-version specific redaction behavior, so a version has to be specified.