Enum ruma::push::FlattenedJsonValue

pub enum FlattenedJsonValue {
    Null,
    Bool(bool),
    Integer(Int),
    String(String),
    Array(Vec<ScalarJsonValue>),
    EmptyObject,
}

Possible JSON values after an object is flattened.

Variants

Null

Represents a null value.

Bool(bool)

Represents a boolean.

Integer(Int)

Represents an integer.

String(String)

Represents a string.

Array(Vec<ScalarJsonValue>)

Represents an array.

EmptyObject

Represents an empty object.

Implementations

impl FlattenedJsonValue

pub fn as_bool(&self) -> Option<bool>

If the FlattenedJsonValue is a Bool, return the inner value.

pub fn as_integer(&self) -> Option<Int>

If the FlattenedJsonValue is an Integer, return the inner value.

pub fn as_str(&self) -> Option<&str>

If the FlattenedJsonValue is a String, return a reference to the inner value.

pub fn as_array(&self) -> Option<&[ScalarJsonValue]>

If the FlattenedJsonValue is an Array, return a reference to the inner value.