Struct ruma::push::Ruleset

#[non_exhaustive]
pub struct Ruleset {
    pub content: IndexSet<PatternedPushRule, RandomState>,
    pub override_: IndexSet<ConditionalPushRule, RandomState>,
    pub room: IndexSet<SimplePushRule, RandomState>,
    pub sender: IndexSet<SimplePushRule, RandomState>,
    pub underride: IndexSet<ConditionalPushRule, RandomState>,
}

A push ruleset scopes a set of rules according to some criteria.

For example, some rules may only be applied for messages from a particular sender, a particular room, or by default. The push ruleset contains the entire set of scopes and rules.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

content: IndexSet<PatternedPushRule, RandomState>

These rules configure behavior for (unencrypted) messages that match certain patterns.

override_: IndexSet<ConditionalPushRule, RandomState>

These user-configured rules are given the highest priority.

This field is named override_ instead of override because the latter is a reserved keyword in Rust.

room: IndexSet<SimplePushRule, RandomState>

These rules change the behavior of all messages for a given room.

sender: IndexSet<SimplePushRule, RandomState>

These rules configure notification behavior for messages from a specific Matrix user ID.

underride: IndexSet<ConditionalPushRule, RandomState>

These rules are identical to override rules, but have a lower priority than content, room and sender rules.

Implementations

impl Ruleset

pub fn server_default(user_id: &UserId) -> Ruleset

The list of all predefined push rules.

Parameters

user_id: the user for which to generate the default rules. Some rules depend on the user’s ID (for instance those to send notifications when they are mentioned).

impl Ruleset

pub fn new() -> Ruleset

Creates an empty Ruleset.

pub fn iter(&self) -> RulesetIter<'_>

Creates a borrowing iterator over all push rules in this Ruleset.

For an owning iterator, use .into_iter().

pub fn add(&mut self, rule: AnyPushRule) -> bool

Adds a rule to the rule set.

Returns true if the new rule was correctly added, and false if a rule with the same rule_id is already present for this kind of rule.

pub fn get_match<T>(
&self,
event: &Raw<T>,
context: &PushConditionRoomCtx
) -> Option<AnyPushRuleRef<'_>>

Get the first push rule that applies to this event, if any.

Arguments

event - The raw JSON of a room message event.

context - The context of the message and room at the time of the event.

pub fn get_actions<T>(
&self,
event: &Raw<T>,
context: &PushConditionRoomCtx
) -> &[Action]

Get the push actions that apply to this event.

Returns an empty slice if no push rule applies.

Arguments

event - The raw JSON of a room message event.

context - The context of the message and room at the time of the event.