Struct ruma::serde::Raw

pub struct Raw<T> { /* private fields */ }

A wrapper around Box<RawValue>, to be used in place of any type in the Matrix endpoint definition to allow request and response types to contain that said type represented by the generic argument Ev.

Ruma offers the Raw wrapper to enable passing around JSON text that is only partially validated. This is useful when a client receives events that do not follow the spec perfectly or a server needs to generate reference hashes with the original canonical JSON string. All event structs and enums implement Serialize / Deserialize, Raw should be used to pass around events in a lossless way.

let json = r#"{ "type": "imagine a full event", "content": {...} }"#;

let deser = serde_json::from_str::<Raw<AnyTimelineEvent>>(json)
    .unwrap() // the first Result from serde_json::from_str, will not fail
    .deserialize() // deserialize to the inner type
    .unwrap(); // finally get to the AnyTimelineEvent

Implementations

impl<T> Raw<T>

pub fn new(val: &T) -> Result<Raw<T>, Error>
where T: Serialize,

Create a Raw by serializing the given T.

Shorthand for serde_json::value::to_raw_value(val).map(Raw::from_json), but specialized to T.

Errors

Fails if Ts Serialize implementation fails.

pub fn from_json(json: Box<RawValue>) -> Raw<T>

Create a Raw from a boxed RawValue.

pub fn from_json_string(json: String) -> Result<Raw<T>, Error>

Convert an owned String of JSON data to Raw<T>.

This function is equivalent to serde_json::from_str::<Raw<T>> except that an allocation and copy is avoided if both of the following are true:

the input has no leading or trailing whitespace, and

the input has capacity equal to its length.

pub fn json(&self) -> &RawValue

Access the underlying json value.

pub fn into_json(self) -> Box<RawValue>

Convert self into the underlying json value.

pub fn get_field<'a, U>(&'a self, field_name: &str) -> Result<Option<U>, Error>
where U: Deserialize<'a>,

Try to access a given field inside this Raw, assuming it contains an object.

Returns Err(_) when the contained value is not an object, or the field exists but is fails to deserialize to the expected type.

Returns Ok(None) when the field doesn’t exist or is null.

Example

if raw_event.get_field::<String>("type")?.as_deref() == Some("org.custom.matrix.event") {
    let event = raw_event.deserialize_as::<CustomMatrixEvent>()?;
    // ...
}

pub fn deserialize<'a>(&'a self) -> Result<T, Error>
where T: Deserialize<'a>,

Try to deserialize the JSON as the expected type.

pub fn deserialize_as<'a, U>(&'a self) -> Result<U, Error>
where U: Deserialize<'a>,

Try to deserialize the JSON as a custom type.

pub fn cast<U>(self) -> Raw<U>

Turns Raw<T> into Raw<U> without changing the underlying JSON.

This is useful for turning raw specific event types into raw event enum types.

pub fn cast_ref<U>(&self) -> &Raw<U>

Turns &Raw<T> into &Raw<U> without changing the underlying JSON.

This is useful for turning raw specific event types into raw event enum types.