Function ruma::serde::time::opt_s_since_unix_epoch::serialize

pub fn serialize<S>(
    opt_time: &Option<SystemTime>, 
    serializer: S
) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error> where
    S: Serializer, 

Serialize an Option<SystemTime>.

Will fail if integer is greater than the maximum integer that can be unambiguously represented by an f64.