Function ruma::signatures::content_hash

pub fn content_hash(object: &BTreeMap<String, CanonicalJsonValue>) -> String

This is supported on crate feature signatures only.

Creates a content hash for an event.

Returns the hash as a Base64-encoded string, using the standard character set, without padding.

The content hash of an event covers the complete event including the unredacted contents. It is used during federation and is described in the Matrix server-server specification.

Parameters

object: A JSON object to generate a content hash for.