Struct ruma::signatures::Signature

pub struct Signature {
    pub algorithm: Algorithm,
    pub signature: Vec<u8, Global>,
    pub version: String,
}

This is supported on crate feature signatures only.

A digital signature.

Fields
algorithm: Algorithm

The cryptographic algorithm that generated this signature.

signature: Vec<u8, Global>

The signature data.

version: String

The “version” of the key identifier for the public key used to generate this signature.

Implementations

impl Signature

pub fn new(id: &str, bytes: &[u8]) -> Result<Signature, Error>

Creates a signature from raw bytes.

While a signature can be created directly using struct literal syntax, this constructor can be used to automatically determine the algorithm and version from a key identifier in the form algorithm:version, e.g. “ed25519:1”.

This constructor will ensure that the version does not contain characters that violate the guidelines in the specification. Because it may be necessary to represent signatures with versions that don’t adhere to these guidelines, it’s possible to simply use the struct literal syntax to construct a Signature with an arbitrary key.

Parameters

id: A key identifier, e.g. “ed25519:1”.

bytes: The digital signature, as a series of bytes.

Errors

Returns an error if:

The key ID specifies an unknown algorithm.

The key ID is malformed.

The key ID contains a version with invalid characters.

pub fn algorithm(&self) -> &Algorithm

The algorithm used to generate the signature.

pub fn as_bytes(&self) -> &[u8]

The raw bytes of the signature.

pub fn base64(&self) -> String

A Base64 encoding of the signature.

Uses the standard character set with no padding.

pub fn id(&self) -> String

The key identifier, a string containing the signature algorithm and the key “version” separated by a colon, e.g. “ed25519:1”.

pub fn version(&self) -> &str

The “version” of the key used for this signature.

Versions are used as an identifier to distinguish signatures generated from different keys but using the same algorithm on the same homeserver.