Crate ruma::state_res

This is supported on crate feature state-res only.

Modules

event_auth

room_version

Structs

StateResolution

Enums

Error
Represents the various errors that arise when resolving state.

Traits

Event
Abstraction of a PDU so users can have their own PDU types.

Functions

auth_check
Authenticate the incoming event. The steps of authentication are:

auth_types_for_event
For the given event kind what are the relevant auth events that are needed to authenticate this content.

is_power_event

is_power_event_id

is_type_and_key

Type Definitions

EventMap
A mapping of EventId to T, usually a ServerPdu.

Result
Result type for state resolution.

StateMap
A mapping of event type and state_key to some value T, usually an EventId.