Trait ruma::state_res::Event

pub trait Event {

    fn event_id(&self) -> &EventId;

    fn room_id(&self) -> &RoomId;

    fn sender(&self) -> &UserId;

    fn origin_server_ts(&self) -> SystemTime;

    fn kind(&self) -> EventType;

    fn content(&self) -> Value;

    fn state_key(&self) -> Option<String>;

    fn prev_events(&self) -> Vec<EventId, Global>;

    fn depth(&self) -> &UInt;

    fn auth_events(&self) -> Vec<EventId, Global>;

    fn redacts(&self) -> Option<&EventId>;

    fn unsigned(&self) -> &BTreeMap<String, Value>;

    fn hashes(&self) -> &EventHash;

    fn signatures(
        &self
    ) -> BTreeMap<Box<ServerName, Global>, BTreeMap<KeyId<SigningKeyAlgorithm, KeyName>, String>>;

}

This is supported on crate feature state-res only.

Abstraction of a PDU so users can have their own PDU types.

Required methods

fn event_id(&self) -> &EventId

The EventId of this event.

fn room_id(&self) -> &RoomId

The RoomId of this event.

fn sender(&self) -> &UserId

The UserId of this event.

fn origin_server_ts(&self) -> SystemTime

The time of creation on the originating server.

fn kind(&self) -> EventType

The kind of event.

fn content(&self) -> Value

The UserId of this PDU.

fn state_key(&self) -> Option<String>

The state key for this event.

fn prev_events(&self) -> Vec<EventId, Global>

The events before this event.

fn depth(&self) -> &UInt

The maximum number of prev_events plus 1.

This is only used in state resolution version 1.

fn auth_events(&self) -> Vec<EventId, Global>

All the authenticating events for this event.

fn redacts(&self) -> Option<&EventId>

If this event is a redaction event this is the event it redacts.

fn unsigned(&self) -> &BTreeMap<String, Value>

The unsigned content of this event.

fn hashes(&self) -> &EventHash

fn signatures(
&self
) -> BTreeMap<Box<ServerName, Global>, BTreeMap<KeyId<SigningKeyAlgorithm, KeyName>, String>>

Loading content...