Struct ruma::ClientSecret

pub struct ClientSecret(/* private fields */);

A client secret.

Client secrets in Matrix are opaque character sequences of [0-9a-zA-Z.=_-]. Their length must must not exceed 255 characters.

You can create one from a string (using ClientSecret::parse()) but the recommended way is to use ClientSecret::new() to generate a random one. If that function is not available for you, you need to activate this crate’s rand Cargo feature.

Implementations

impl ClientSecret

pub fn as_str(&self) -> &str

Creates a string slice from this ClientSecret.

pub fn as_bytes(&self) -> &[u8]

Creates a byte slice from this ClientSecret.

impl ClientSecret

pub fn parse(s: impl AsRef<str>) -> Result<OwnedClientSecret, Error>

Try parsing a &str into an OwnedClientSecret.

The same can also be done using FromStr, TryFrom or TryInto. This function is simply more constrained and thus useful in generic contexts.

pub fn parse_box( s: impl AsRef<str> + Into<Box<str>> ) -> Result<Box<ClientSecret>, Error>

Try parsing a &str into a Box<ClientSecret>.

The same can also be done using FromStr, TryFrom or TryInto. This function is simply more constrained and thus useful in generic contexts.

pub fn parse_rc( s: impl AsRef<str> + Into<Rc<str>> ) -> Result<Rc<ClientSecret>, Error>

Try parsing a &str into an Rc<ClientSecret>.

pub fn parse_arc( s: impl AsRef<str> + Into<Arc<str>> ) -> Result<Arc<ClientSecret>, Error>

Try parsing a &str into an Arc<ClientSecret>.

impl ClientSecret

pub fn new() -> OwnedClientSecret

Creates a random client secret.

This will currently be a UUID without hyphens, but no guarantees are made about the structure of client secrets generated from this function.