Struct ruma::UserId

pub struct UserId(/* private fields */);

A Matrix user ID.

A UserId is generated randomly or converted from a string slice, and can be converted back into a string as needed.

assert_eq!(<&UserId>::try_from("@carl:example.com").unwrap(), "@carl:example.com");

Implementations

impl UserId

pub fn as_str(&self) -> &str

Creates a string slice from this UserId.

pub fn as_bytes(&self) -> &[u8]

Creates a byte slice from this UserId.

impl UserId

pub fn parse(s: impl AsRef<str>) -> Result<OwnedUserId, Error>

Try parsing a &str into an OwnedUserId.

The same can also be done using FromStr, TryFrom or TryInto. This function is simply more constrained and thus useful in generic contexts.

pub fn parse_box( s: impl AsRef<str> + Into<Box<str>> ) -> Result<Box<UserId>, Error>

Try parsing a &str into a Box<UserId>.

The same can also be done using FromStr, TryFrom or TryInto. This function is simply more constrained and thus useful in generic contexts.

pub fn parse_rc(s: impl AsRef<str> + Into<Rc<str>>) -> Result<Rc<UserId>, Error>

Try parsing a &str into an Rc<UserId>.

pub fn parse_arc( s: impl AsRef<str> + Into<Arc<str>> ) -> Result<Arc<UserId>, Error>

Try parsing a &str into an Arc<UserId>.

impl UserId

pub fn new(server_name: &ServerName) -> OwnedUserId

Attempts to generate a UserId for the given origin server with a localpart consisting of 12 random ASCII characters.

pub fn parse_with_server_name( id: impl AsRef<str> + Into<Box<str>>, server_name: &ServerName ) -> Result<OwnedUserId, Error>

Attempts to complete a user ID, by adding the colon + server name and @ prefix, if not present already.

This is a convenience function for the login API, where a user can supply either their full user ID or just the localpart. It only supports a valid user ID or a valid user ID localpart, not the localpart plus the @ prefix, or the localpart plus server name without the @ prefix.

pub fn parse_with_server_name_rc( id: impl AsRef<str> + Into<Rc<str>>, server_name: &ServerName ) -> Result<Rc<UserId>, Error>

Variation of parse_with_server_name that returns Rc<Self>.

pub fn parse_with_server_name_arc( id: impl AsRef<str> + Into<Arc<str>>, server_name: &ServerName ) -> Result<Arc<UserId>, Error>

Variation of parse_with_server_name that returns Arc<Self>.

pub fn localpart(&self) -> &str

Returns the user’s localpart.

pub fn server_name(&self) -> &ServerName

Returns the server name of the user ID.

pub fn is_historical(&self) -> bool

Whether this user ID is a historical one.

A historical user ID is one that doesn’t conform to the latest specification of the user ID grammar but is still accepted because it was previously allowed.

pub fn matrix_to_uri(&self) -> MatrixToUri

Create a matrix.to URI for this user ID.

Example

use ruma_common::user_id;

let message = format!(
    r#"Thanks for the update <a href="{link}">{display_name}</a>."#,
    link = user_id!("@jplatte:notareal.hs").matrix_to_uri(),
    display_name = "jplatte",
);

pub fn matrix_uri(&self, chat: bool) -> MatrixUri

Create a matrix: URI for this user ID.

If chat is true, a click on the URI should start a direct message with the user.

Example

use ruma_common::user_id;

let message = format!(
    r#"Thanks for the update <a href="{link}">{display_name}</a>."#,
    link = user_id!("@jplatte:notareal.hs").matrix_uri(false),
    display_name = "jplatte",
);