Crate ruma

Types and traits for working with the Matrix protocol.

This crate re-exports things from all of the other ruma crates so you don’t have to manually keep all the versions in sync.

Which crates are re-exported can be configured through cargo features.

⚠ Some details might be missing because rustdoc has trouble with re-exports so you may need to refer to other crates’ documentations.

API features

Depending on which parts of Matrix are relevant to you, activate the following features:

appservice-api – Application Service API.

client-api – Client-Server API.

federation-api – Server-Server (Federation) API.

identity-service-api – Identity Service API.

push-gateway-api – Push Gateway API.

These features have client- and server-optimized variants that are enabled respectively with the -c and -s suffixes. For example:

client-api-c – The Client-Server API optimized for the client side.

client-api-s – The Client-Server API optimized for the server side.

Compatibility feature

compat increases compatibility with other parts of the Matrix ecosystem, at the expense of deviating from the specification.

Convenience features

These features are only useful if you want to use a method that requires it:

either

rand

markdown

Unstable features

By using these features, you opt out of all semver guarantees Ruma otherwise provides:

unstable-exhaustive-types – Most types in Ruma are marked as non-exhaustive to avoid breaking changes when new fields are added in the specification. This feature compiles all types as exhaustive.

unstable-mscXXXX, where XXXX is the MSC number – Upcoming Matrix features that may be subject to change or removal.

unstable-pre-spec – Undocumented Matrix features that may be subject to change or removal.

Common features

These submodules are usually activated by the API features when needed:

api

events

signatures

ruma-client features

The client feature activates [ruma::client][client], and client-ext-client-api activates ruma-clients client-api feature. All other client-* features activate the same feature without the client- prefix on ruma-client. See the crate’s documentation for the effect of these features.

If you are viewing this on docs.rs, you can have a look at the feature dependencies by clicking Feature flags in the toolbar at the top.

Re-exports

pub use assign::assign;

pub use js_int::int;

pub use js_int::uint;

pub use js_int::Int;

pub use js_int::UInt;

Modules

api

(De)serializable types for various Matrix APIs requests and responses and abstractions for them.

authentication

Common types for authentication.

directory

Common types for room directory endpoints.

encryption

Common types for encryption related tasks.

events

(De)serializable types for the events in the Matrix specification. These types are used by other Ruma crates.

matrix_uri

Matrix URIs.

power_levels

Common types for the m.room.power_levels event.

presence

Common types for the presence module.

push

Common types for the push notifications module.

receipt

Common types for receipts.

room

Common types for rooms.

serde

(De)serialization helpers for other Ruma crates.

signatures

Digital signatures according to the Matrix specification.

thirdparty

Common types for the third party networks module.

to_device

Common types for the Send-To-Device Messaging

user_id

Matrix user identifiers.

Macros

device_id

Shorthand for <&DeviceId>::from.

device_key_id

Compile-time checked DeviceKeyId construction.

event_id

Compile-time checked EventId construction.

mxc_uri

Compile-time checked MxcUri construction.

room_alias_id

Compile-time checked RoomAliasId construction.

room_id

Compile-time checked RoomId construction.

room_version_id

Compile-time checked RoomVersionId construction.

server_name

Compile-time checked ServerName construction.

server_signing_key_id

Compile-time checked ServerSigningKeyId construction.

user_id

Compile-time checked UserId construction.

Structs

ClientSecret

A client secret.

DeviceId

A Matrix key ID.

DeviceKeyId

A key algorithm and a device id, combined with a ‘:’.

EventId

A Matrix event ID.

KeyId

A key algorithm and key name delimited by a colon.

KeyName

A Matrix key identifier.

MatrixToUri

The matrix.to URI representation of a user, room or event.

MatrixUri

The matrix: URI representation of a user, room or event.

MilliSecondsSinceUnixEpoch

A timestamp represented as the number of milliseconds since the unix epoch.

MxcUri

A URI that should be a Matrix-spec compliant MXC URI.

OwnedClientSecret

Owned variant of ClientSecret

OwnedDeviceId

Owned variant of DeviceId

OwnedDeviceKeyId

Owned variant of DeviceKeyId

OwnedEventId

Owned variant of EventId

OwnedKeyId

Owned variant of KeyId

OwnedKeyName

Owned variant of KeyName

OwnedMxcUri

Owned variant of MxcUri

OwnedRoomAliasId

Owned variant of RoomAliasId

OwnedRoomId

Owned variant of RoomId

OwnedRoomName

Owned variant of RoomName

OwnedRoomOrAliasId

Owned variant of RoomOrAliasId

OwnedServerName

Owned variant of ServerName

OwnedSessionId

Owned variant of SessionId

OwnedTransactionId

Owned variant of TransactionId

OwnedUserId

Owned variant of UserId

RoomAliasId

A Matrix room alias ID.

RoomId

A Matrix room ID.

RoomName

The name of a room.

RoomOrAliasId

A Matrix room ID or a Matrix room alias ID.

SecondsSinceUnixEpoch

A timestamp represented as the number of seconds since the unix epoch.

ServerName

A Matrix-spec compliant server name.

SessionId

A session ID.

Signatures

Map of all signatures, grouped by entity

TransactionId

A Matrix transaction ID.

UserId

A Matrix user ID.

Enums

DeviceKeyAlgorithm

The basic key algorithms in the specification.

EventEncryptionAlgorithm

An encryption algorithm to be used to encrypt messages sent to a room.

IdParseError

An error encountered when trying to parse an invalid ID string.

RoomVersionId

A Matrix room version ID.

SigningKeyAlgorithm

The signing key algorithms defined in the Matrix spec.

Type Definitions

DeviceSignatures

Map of device signatures for an event, grouped by user.

DeviceSigningKeyId

Algorithm + key name for device keys.

EntitySignatures

Map of key identifier to signature values.

OwnedDeviceSigningKeyId

Algorithm + key name for device keys.

OwnedServerSigningKeyId

Algorithm + key name for homeserver signing keys.

OwnedSigningKeyId

Algorithm + key name for signing keys.

ServerSignatures

Map of server signatures for an event, grouped by server.

ServerSigningKeyId

Algorithm + key name for homeserver signing keys.

Derive Macros

Incoming

Generating an ‘Incoming’ version of the type this derive macro is used on.