Crate ruma::api::push_gateway

This is supported on crate feature api and (crate features push-gateway-api or push-gateway-api-c or push-gateway-api-s) only.

(De)serializable types for the Matrix Push Gateway API. These types can be shared by push gateway and server code.

Modules

send_event_notification
Endpoint to notify a push gateway about an event or update the number of unread notifications a user has.