Trait ruma::api::IncomingRequest

pub trait IncomingRequest {
    type EndpointError: EndpointError;
    type OutgoingResponse: TryInto<Response<Vec<u8, Global>>>;

    pub const METADATA: Metadata;

    pub fn try_from_http_request(
        req: Request<Vec<u8, Global>>
    ) -> Result<Self, FromHttpRequestError>;
}

This is supported on crate feature api only.

A request type for a Matrix API endpoint, used for receiving requests.

Associated Types

type EndpointError: EndpointError

A type capturing the error conditions that can be returned in the response.

type OutgoingResponse: TryInto<Response<Vec<u8, Global>>>

Response type to return when the request is successful.

Loading content...
Associated Constants

pub const METADATA: Metadata

Metadata about the endpoint.

Loading content...
Required methods

pub fn try_from_http_request(
req: Request<Vec<u8, Global>>
) -> Result<Self, FromHttpRequestError>

Tries to turn the given http::Request into this request type.

Loading content...