Trait ruma::events::RedactContent

pub trait RedactContent {
    type Redacted;
    fn redact(self, version: &RoomVersionId) -> Self::Redacted;
}

This is supported on crate feature events only.

Trait to define the behavior of redact an event’s content object.

Associated Types

type Redacted

The redacted form of the event’s content.

Required methods

fn redact(self, version: &RoomVersionId) -> Self::Redacted

Transform self into a redacted form (removing most or all fields) according to the spec.

A small number of events have room-version specific redaction behavior, so a version has to be specified.

Where applicable, it is prefered to use Redact::redact on the outer event.