Module ruma::push

Common types for the push notifications module.

Understanding the types of this module

Push rules are grouped in RuleSets, and are grouped in five kinds (for more details about the different kind of rules, see the Ruleset documentation, or the specification). These five kinds are, by order of priority:

override rules

content rules

room rules

sender rules

underride rules

Structs

ConditionalPushRule

Like SimplePushRule, but with an additional conditions field.

ConditionalPushRuleInit

Initial set of fields of ConditionalPushRule.

FlattenedJson

The flattened representation of a JSON object.

PatternedPushRule

Like SimplePushRule, but with an additional pattern field.

PatternedPushRuleInit

Initial set of fields of PatterenedPushRule.

PushConditionRoomCtx

The context of the room associated to an event to be able to test all push conditions.

PusherData

Information for the pusher implementation itself.

RoomMemberCountIs

A decimal integer optionally prefixed by one of ==, <, >, >= or <=.

Ruleset

A push ruleset scopes a set of rules according to some criteria.

RulesetIntoIter

Iterator type for Ruleset

RulesetIter

Iterator type for Ruleset

SimplePushRule

A push rule is a single rule that states under what conditions an event should be passed onto a push gateway and how the notification should be presented.

SimplePushRuleInit

Initial set of fields of SimplePushRule.

Enums

Action

This represents the different actions that should be taken when a rule is matched, and controls how notifications are delivered to the client.

AnyPushRule

The kinds of push rules that are available.

AnyPushRuleRef

Reference to any kind of push rule.

ComparisonOperator

One of ==, <, >, >= or <=.

PushCondition

A condition that must apply for an associated push rule’s action to be taken.

PushFormat

A special format that the homeserver should use when sending notifications to a Push Gateway. Currently, only “event_id_only” is supported as of Push Gateway API r0.1.1.

Tweak

The set_tweak action.