Function ruma::state_res::lexicographical_topological_sort

pub fn lexicographical_topological_sort<F>(
    graph: &HashMap<EventId, HashSet<EventId, RandomState>, RandomState>, 
    key_fn: F
) -> Result<Vec<EventId, Global>, Error> where
    F: Fn(&EventId) -> Result<(Int, MilliSecondsSinceUnixEpoch, EventId), Error>, 

Sorts the event graph based on number of outgoing/incoming edges.

key_fn is used as a tie breaker. The tie breaker happens based on power level, age, and event_id.